MaxRetransmits を W3C の仕様に合わせて ushort に修正

+++ Sora/Sora.cs
@@ -57,7 +57,7 @@ public class Sora : IDisposable
         // 以下は null でない場合のみ設定される
         public bool? Ordered;
         public int? MaxPacketLifeTime;
-        public int? MaxRetransmits;
+        public ushort? MaxRetransmits;
         public string Protocol;
         public bool? Compress;
     }
